requirements.txt
torch==2.2.0
transformers==4.38.2
tokenizers==0.14.1
huggingface-hub==0.22.2
librosa==0.9.2
decord==0.6.0
numpy==1.26.4
opencv-python==4.8.0.74
tqdm==4.65.0
scikit-learn==1.3.0
sentence-transformers==2.6.1
ffmpeg-python==0.2.0
